+++ plugins/minister/src/minister/__init__.py
@@ -14,4 +14,4 @@ __all__ = [
     "run_cli",
 ]
 
-__version__ = "1.8.2"
+__version__ = "1.8.3"